+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroExtensionRegistrarConfigurator.kt
@@ -67,6 +67,8 @@ class MetroExtensionRegistrarConfigurator(testServices: TestServices) :
           MetroDirectives.ENABLE_FULL_BINDING_GRAPH_VALIDATION in module.directives
         enableGraphImplClassAsReturnType =
           MetroDirectives.ENABLE_GRAPH_IMPL_CLASS_AS_RETURN_TYPE in module.directives
+        generateContributionHints =
+          module.directives.singleOrZeroValue(MetroDirectives.GENERATE_CONTRIBUTION_HINTS) ?: true
         generateContributionHintsInFir =
           MetroDirectives.GENERATE_CONTRIBUTION_HINTS_IN_FIR in module.directives
         if (transformProvidersToPrivate) {